% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_coins}
\alias{get_coins}
\title{List all available coins.}
\usage{
get_coins(sign = FALSE, app_name = NULL)
}
\arguments{
\item{sign}{logical. Should the server sign the response? Defaults to FALSE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing details for each coin available through the API.
}
\description{
\code{get_coins} provides general information about all coins available through
the CryptoCompare API.
}
\examples{
\dontrun{
coins <- get_coins()
head(coins)
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
